[RTD-2149] fix native bug
